# integrated outliers page to selection page
