+++ data_sweep/lib/trial_codes/all_issues.csv
@@ -1,11 +1,11 @@
 Name,Age,Gender,Description,Date
 Alice,23,Female,Loves painting,2023-01-01
-Bob,29999,M,Software engineer,
-,8,Male,Avid reader,2023-02-15
+Bob,29999,M,Software engineer,2023-01-01
+Ace,8,Male,Avid reader,2023-02-15
 Diana,32,Woman,,2023/03/10
-Evan,0.0001,F,Fitness enthusiast,"March 20, 2023"
+Evan,0.0001,F,Fitness enthusiast,2023-01-01
 Fiona,27,Female,,2023-04-25
-,30,M,,
+Vivs,30,M,,2023-01-01
 Hannah,1080,Female,Book lover,2023-05-30
-Imani,1000,Male,Traveler,2023-06-01
-James,0.00002,Male,,
+Imani,1,Male,Traveler,2023-06-01
+James,3,Male,,2023-01-01
